requirements.txt
# =================================================================
# Author: Awais Nazeer (ZRR Gujjar)
# Email: awaisnazeer07@gmail.com
# =================================================================
# Bakery OCR System - Python Dependencies
# =================================================================

flask==2.2.3
flask-cors==3.0.10
pytesseract==0.3.10
opencv-python==4.7.0.72
pandas==1.5.3
numpy==1.24.2
Pillow==9.5.0
gunicorn==20.1.0 